requirements/local.txt
-r base.txt
django-debug-toolbar==1.4
django-extensions==1.3.3
django-session-security==2.3.1
Fabric==1.8.2
mixer==3.0.0
testfixtures==3.0.1
Werkzeug==0.9.4
